The dataset.
This dataset was used for this model. 
This dataset provides a comprehensive collection of features related to houses in California, with the primary aim of facilitating the prediction of house rent prices. It includes 80 columns and 1460 rows, offering a rich set of information for model training and evaluation.
Target Variable:
The dataset aims to predict the house rent prices, making it suitable for regression models. The 'SalePrice' column can be used as the target variable for training and evaluating predictive models.

Columns:

Id: Unique identifier for each record.
MSSubClass: The building class
MSZoning: The general zoning classification of the property.
LotFrontage: Linear feet of street connected to property.
LotArea: Lot size in square feet.
Street: Type of road access to property.
Alley: Type of alley access to property.
LotShape: General shape of the property.
LandContour: Flatness of the property.
… (and many more)
Use Case:
Ideal for exploring and implementing regression models, particularly Linear Regression, to predict house rent prices based on various features associated with the properties.

Dataset Size:
80 columns
1460 rows